move bundle code inside its own directory
