+++ src/aliceVision/sfm/bundle/BundleAdjustment.hpp
