+++ src/aliceVision/sfm/bundle/BundleAdjustmentCeres.cpp
@@ -5,27 +5,21 @@
 // v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include <aliceVision/sfm/BundleAdjustmentSymbolicCeres.hpp>
-
-#include <aliceVision/alicevision_omp.hpp>
-
+#include <aliceVision/sfm/bundle/BundleAdjustmentCeres.hpp>
+#include <aliceVision/sfm/ResidualErrorFunctor.hpp>
+#include <aliceVision/sfm/ResidualErrorConstraintFunctor.hpp>
+#include <aliceVision/sfm/ResidualErrorRotationPriorFunctor.hpp>
 #include <aliceVision/sfmData/SfMData.hpp>
+#include <aliceVision/alicevision_omp.hpp>
 #include <aliceVision/config.hpp>
-#include <aliceVision/camera/Equidistant.hpp>
+#include <aliceVision/camera/camera.hpp>
+
 #include <boost/filesystem.hpp>
 
-
-#include <aliceVision/sfm/manifolds/so3.hpp>
-#include <aliceVision/sfm/manifolds/se3.hpp>
-#include <aliceVision/sfm/manifolds/intrinsics.hpp>
-
-#include <aliceVision/sfm/costfunctions/projection.hpp>
-#include <aliceVision/sfm/costfunctions/projectionSimple.hpp>
-#include <aliceVision/sfm/costfunctions/panoramaEquidistant.hpp>
-#include <aliceVision/sfm/costfunctions/panoramaPinhole.hpp>
-#include <aliceVision/sfm/costfunctions/rotationPrior.hpp>
+#include <ceres/rotation.h>
 
 #include <fstream>
+#include <memory>
 
 
 namespace fs = boost::filesystem;
@@ -36,55 +30,327 @@ namespace sfm {
 using namespace aliceVision::camera;
 using namespace aliceVision::geometry;
 
+class IntrinsicsManifold : public ceres::Manifold {
+ public:
+  explicit IntrinsicsManifold(size_t parametersSize, double focalRatio, bool lockFocal, bool lockFocalRatio, bool lockCenter, bool lockDistortion)
+  : _ambientSize(parametersSize),
+    _focalRatio(focalRatio),
+    _lockFocal(lockFocal),
+    _lockFocalRatio(lockFocalRatio),
+    _lockCenter(lockCenter),
+    _lockDistortion(lockDistortion)
+  {
+    _distortionSize = _ambientSize - 4;
+    _tangentSize = 0;
 
-void BundleAdjustmentSymbolicCeres::addPose(const sfmData::CameraPose& cameraPose, bool isConstant, SE3::Matrix & poseBlock, ceres::Problem& problem, bool refineTranslation, bool refineRotation)
+    if (!_lockFocal)
+    {
+      if (_lockFocalRatio)
+      {
+        _tangentSize += 1;
+      }
+      else
+      {
+        _tangentSize += 2;
+      }
+    }
+
+    if (!_lockCenter)
+    {
+      _tangentSize += 2;
+    }
+
+    if (!_lockDistortion)
+    {
+      _tangentSize += _distortionSize;
+    }
+  }
+
+  virtual ~IntrinsicsManifold() = default;
+
+
+  bool Plus(const double* x, const double* delta, double* x_plus_delta) const override
+  {
+    for (int i = 0; i < _ambientSize; i++)
+    {
+      x_plus_delta[i] = x[i];
+    }
+    
+    size_t posDelta = 0;
+    if (!_lockFocal)
+    {
+      if (_lockFocalRatio)
+      {
+        x_plus_delta[0] = x[0] + _focalRatio * delta[posDelta]; 
+        x_plus_delta[1] = x[1] + delta[posDelta];
+        posDelta++;
+      }
+      else
+      {
+        x_plus_delta[0] = x[0] + delta[posDelta];
+        posDelta++;
+        x_plus_delta[1] = x[1] + delta[posDelta];
+        posDelta++;
+      }
+    }
+
+    if (!_lockCenter)
+    {
+      x_plus_delta[2] = x[2] + delta[posDelta]; 
+      posDelta++;
+
+      x_plus_delta[3] = x[3] + delta[posDelta];
+      posDelta++;
+    }
+
+    if (!_lockDistortion)
+    {
+      for (int i = 0; i < _distortionSize; i++)
+      {
+        x_plus_delta[4 + i] = x[4 + i] + delta[posDelta];
+        posDelta++;
+      }
+    }
+
+    return true;
+  }
+
+  bool PlusJacobian(const double* x, double* jacobian) const override
+  {    
+    Eigen::Map<Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor>> J(jacobian, AmbientSize(), TangentSize());
+
+    J.fill(0);
+
+    size_t posDelta = 0;
+    if (!_lockFocal)
+    {
+      if (_lockFocalRatio)
+      {
+        J(0, posDelta) = _focalRatio;
+        J(1, posDelta) = 1.0;
+        posDelta++;
+      }
+      else
+      {
+        J(0, posDelta) = 1.0;
+        posDelta++;
+        J(1, posDelta) = 1.0;
+        posDelta++;
+      }
+    }
+
+    if (!_lockCenter)
+    {
+      J(2, posDelta) = 1.0;
+      posDelta++;
+
+      J(3, posDelta) = 1.0;
+      posDelta++;
+    }
+
+    if (!_lockDistortion)
+    {
+      for (int i = 0; i < _distortionSize; i++)
+      {
+        J(4 + i, posDelta) = 1.0;
+        posDelta++;
+      }
+    }
+
+    return true;
+  }
+
+  bool Minus(const double* y, const double* x, double* delta) const override {
+    throw std::invalid_argument("IntrinsicsManifold::Minus() should never be called");
+  }
+
+  bool MinusJacobian(const double* x, double* jacobian) const override {
+    throw std::invalid_argument("IntrinsicsManifold::MinusJacobian() should never be called");
+  }
+
+  int AmbientSize() const override
+  {
+    return _ambientSize;
+  }
+
+  int TangentSize() const override
+  { 
+    return _tangentSize;
+  }
+
+ private:
+  size_t _distortionSize;
+  size_t _ambientSize;
+  size_t _tangentSize;
+  double _focalRatio;
+  bool _lockFocal;
+  bool _lockFocalRatio;
+  bool _lockCenter;
+  bool _lockDistortion;
+};
+
+/**
+ * @brief Create the appropriate cost functor according the provided input camera intrinsic model
+ * @param[in] intrinsicPtr The intrinsic pointer
+ * @param[in] observation The corresponding observation
+ * @return cost functor
+ */
+ceres::CostFunction* createCostFunctionFromIntrinsics(const IntrinsicBase* intrinsicPtr, const sfmData::Observation& observation)
 {
-  const Mat3& R = cameraPose.getTransform().rotation();
-  const Vec3& t = cameraPose.getTransform().translation();
-  
+  int w = intrinsicPtr->w();
+  int h = intrinsicPtr->h();
 
-  poseBlock = SE3::Matrix::Identity();
-  poseBlock.block<3,3>(0, 0) = R;
-  poseBlock.block<3,1>(0, 3) = t;
-  double * poseBlockPtr = poseBlock.data();
-  problem.AddParameterBlock(poseBlockPtr, 16);
-
-
-  // add pose parameter to the all parameters blocks pointers list
-  _allParametersBlocks.push_back(poseBlockPtr);
-
-  // keep the camera extrinsics constants
-  if(cameraPose.isLocked() || isConstant || (!refineTranslation && !refineRotation))
+  // Apply undistortion to observation
+  sfmData::Observation obsUndistorted = observation;
+  const camera::IntrinsicScaleOffsetDisto* intrinsicDistortionPtr =
+    dynamic_cast<const camera::IntrinsicScaleOffsetDisto*>(intrinsicPtr);
+  if (intrinsicDistortionPtr)
   {
-    // set the whole parameter block as constant.
-    _statistics.addState(EParameter::POSE, EParameterState::CONSTANT);
-
-    problem.SetParameterBlockConstant(poseBlockPtr);
-    return;
+    auto undistortion = intrinsicDistortionPtr->getUndistortion();
+    if (undistortion)
+    {
+      obsUndistorted.x = undistortion->undistort(observation.x);
+    }
   }
 
-  if (refineRotation || refineTranslation)
+  switch(intrinsicPtr->getType())
   {
-    problem.SetManifold(poseBlockPtr, new SE3ManifoldLeft(refineRotation, refineTranslation));
+    case EINTRINSIC::PINHOLE_CAMERA:
+      return new ceres::AutoDiffCostFunction<ResidualErrorFunctor_Pinhole, 2, 4, 6, 3>(
+        new ResidualErrorFunctor_Pinhole(w, h, obsUndistorted));
+    case EINTRINSIC::PINHOLE_CAMERA_RADIAL1:
+      return new ceres::AutoDiffCostFunction<ResidualErrorFunctor_PinholeRadialK1, 2, 5, 6, 3>(
+        new ResidualErrorFunctor_PinholeRadialK1(w, h, obsUndistorted));
+    case EINTRINSIC::PINHOLE_CAMERA_RADIAL3:
+      return new ceres::AutoDiffCostFunction<ResidualErrorFunctor_PinholeRadialK3, 2, 7, 6, 3>(
+        new ResidualErrorFunctor_PinholeRadialK3(w, h, obsUndistorted));
+    case EINTRINSIC::PINHOLE_CAMERA_3DERADIAL4:
+      return new ceres::AutoDiffCostFunction<ResidualErrorFunctor_Pinhole3DERadial4, 2, 10, 6, 3>(
+        new ResidualErrorFunctor_Pinhole3DERadial4(w, h, obsUndistorted));
+    case EINTRINSIC::PINHOLE_CAMERA_3DECLASSICLD:
+      return new ceres::AutoDiffCostFunction<ResidualErrorFunctor_Pinhole3DEClassicLD, 2, 9, 6, 3>(
+        new ResidualErrorFunctor_Pinhole3DEClassicLD(w, h, obsUndistorted));
+    case EINTRINSIC::PINHOLE_CAMERA_3DEANAMORPHIC4:
+      return new ceres::AutoDiffCostFunction<ResidualErrorFunctor_Pinhole, 2, 4, 6, 3>(
+        new ResidualErrorFunctor_Pinhole(w, h, obsUndistorted));
+    case EINTRINSIC::PINHOLE_CAMERA_BROWN:
+      return new ceres::AutoDiffCostFunction<ResidualErrorFunctor_PinholeBrownT2, 2, 9, 6, 3>(
+        new ResidualErrorFunctor_PinholeBrownT2(w, h, obsUndistorted));
+    case EINTRINSIC::PINHOLE_CAMERA_FISHEYE:
+      return new ceres::AutoDiffCostFunction<ResidualErrorFunctor_PinholeFisheye, 2, 8, 6, 3>(
+        new ResidualErrorFunctor_PinholeFisheye(w, h, obsUndistorted));
+    case EINTRINSIC::PINHOLE_CAMERA_FISHEYE1:
+      return new ceres::AutoDiffCostFunction<ResidualErrorFunctor_PinholeFisheye1, 2, 5, 6, 3>(
+        new ResidualErrorFunctor_PinholeFisheye1(w, h, obsUndistorted));
+    default:
+      throw std::logic_error("Cannot create cost function, unrecognized intrinsic type in BA.");
   }
-  else
-  {
-    ALICEVISION_THROW_ERROR("BundleAdjustmentSymbolicCeres: Constant extrinsics not supported at this time");
-  }
-
-  _statistics.addState(EParameter::POSE, EParameterState::REFINED);
 }
 
-void BundleAdjustmentSymbolicCeres::CeresOptions::setDenseBA()
+/**
+ * @brief Create the appropriate cost functor according the provided input rig camera intrinsic model
+ * @param[in] intrinsicPtr The intrinsic pointer
+ * @param[in] observation The corresponding observation
+ * @return cost functor
+ */
+ceres::CostFunction* createRigCostFunctionFromIntrinsics(const IntrinsicBase* intrinsicPtr, const sfmData::Observation& observation)
+{
+  int w = intrinsicPtr->w();
+  int h = intrinsicPtr->h();
+
+  // Apply undistortion to observation
+  sfmData::Observation obsUndistorted = observation;
+  const camera::IntrinsicScaleOffsetDisto* intrinsicDistortionPtr =
+    dynamic_cast<const camera::IntrinsicScaleOffsetDisto*>(intrinsicPtr);
+  if (intrinsicDistortionPtr)
+  {
+    auto undistortion = intrinsicDistortionPtr->getUndistortion();
+    if (undistortion)
+    {
+      obsUndistorted.x = undistortion->undistort(observation.x);
+    }
+  }
+
+  switch(intrinsicPtr->getType())
+  {
+    case EINTRINSIC::PINHOLE_CAMERA:
+      return new ceres::AutoDiffCostFunction<ResidualErrorFunctor_Pinhole, 2, 4, 6, 6, 3>(
+        new ResidualErrorFunctor_Pinhole(w, h, obsUndistorted));
+    case EINTRINSIC::PINHOLE_CAMERA_RADIAL1:
+      return new ceres::AutoDiffCostFunction<ResidualErrorFunctor_PinholeRadialK1, 2, 5, 6, 6, 3>(
+        new ResidualErrorFunctor_PinholeRadialK1(w, h, obsUndistorted));
+    case EINTRINSIC::PINHOLE_CAMERA_RADIAL3:
+      return new ceres::AutoDiffCostFunction<ResidualErrorFunctor_PinholeRadialK3, 2, 7, 6, 6, 3>(
+        new ResidualErrorFunctor_PinholeRadialK3(w, h, obsUndistorted));
+    case EINTRINSIC::PINHOLE_CAMERA_3DERADIAL4:
+      return new ceres::AutoDiffCostFunction<ResidualErrorFunctor_Pinhole3DERadial4, 2, 10, 6, 6, 3>(
+        new ResidualErrorFunctor_Pinhole3DERadial4(w, h, obsUndistorted));
+    case EINTRINSIC::PINHOLE_CAMERA_3DECLASSICLD:
+      return new ceres::AutoDiffCostFunction<ResidualErrorFunctor_Pinhole3DEClassicLD, 2, 9, 6, 6, 3>(
+        new ResidualErrorFunctor_Pinhole3DEClassicLD(w, h, obsUndistorted));
+    case EINTRINSIC::PINHOLE_CAMERA_3DEANAMORPHIC4:
+      return new ceres::AutoDiffCostFunction<ResidualErrorFunctor_Pinhole3DEAnamorphic4, 2, 18, 6, 6, 3>(
+        new ResidualErrorFunctor_Pinhole3DEAnamorphic4(w, h, obsUndistorted));
+    case EINTRINSIC::PINHOLE_CAMERA_BROWN:
+      return new ceres::AutoDiffCostFunction<ResidualErrorFunctor_PinholeBrownT2, 2, 9, 6, 6, 3>(
+        new ResidualErrorFunctor_PinholeBrownT2(w, h, obsUndistorted));
+    case EINTRINSIC::PINHOLE_CAMERA_FISHEYE:
+      return new ceres::AutoDiffCostFunction<ResidualErrorFunctor_PinholeFisheye, 2, 8, 6, 6, 3>(
+        new ResidualErrorFunctor_PinholeFisheye(w, h, obsUndistorted));
+    case EINTRINSIC::PINHOLE_CAMERA_FISHEYE1:
+      return new ceres::AutoDiffCostFunction<ResidualErrorFunctor_PinholeFisheye1, 2, 5, 6, 6, 3>(
+        new ResidualErrorFunctor_PinholeFisheye1(w, h, obsUndistorted));
+    default:
+      throw std::logic_error("Cannot create rig cost function, unrecognized intrinsic type in BA.");
+  }
+}
+
+/**
+ * @brief Create the appropriate cost functor according the provided input camera intrinsic model
+ * @param[in] intrinsicPtr The intrinsic pointer
+ * @param[in] observation The corresponding observation
+ * @return cost functor
+ */
+ceres::CostFunction* createConstraintsCostFunctionFromIntrinsics(const IntrinsicBase* intrinsicPtr, const Vec2& observation_first, const Vec2& observation_second)
+{
+  double radius = 0.0;
+  const camera::Equidistant * equi = dynamic_cast<const camera::Equidistant *>(intrinsicPtr);
+  if (equi) {
+    radius = equi->getCircleRadius();
+  }
+
+  int w = intrinsicPtr->w();
+  int h = intrinsicPtr->h();
+
+  switch(intrinsicPtr->getType())
+  {
+    case EINTRINSIC::PINHOLE_CAMERA:
+      return new ceres::AutoDiffCostFunction<ResidualErrorConstraintFunctor_Pinhole, 2, 3, 6, 6>(new ResidualErrorConstraintFunctor_Pinhole(w, h, observation_first.homogeneous(), observation_second.homogeneous()));
+    case EINTRINSIC::PINHOLE_CAMERA_RADIAL1:
+      return new ceres::AutoDiffCostFunction<ResidualErrorConstraintFunctor_PinholeRadialK1, 2, 4, 6, 6>(new ResidualErrorConstraintFunctor_PinholeRadialK1(w, h, observation_first.homogeneous(), observation_second.homogeneous()));
+    case EINTRINSIC::PINHOLE_CAMERA_RADIAL3:
+      return new ceres::AutoDiffCostFunction<ResidualErrorConstraintFunctor_PinholeRadialK3, 2, 6, 6, 6>(new ResidualErrorConstraintFunctor_PinholeRadialK3(w, h, observation_first.homogeneous(), observation_second.homogeneous()));
+    case EINTRINSIC::PINHOLE_CAMERA_FISHEYE:
+      return new ceres::AutoDiffCostFunction<ResidualErrorConstraintFunctor_PinholeFisheye, 2, 7, 6, 6>(new ResidualErrorConstraintFunctor_PinholeFisheye(w, h, observation_first.homogeneous(), observation_second.homogeneous()));
+    case EQUIDISTANT_CAMERA:
+      return new ceres::AutoDiffCostFunction<ResidualErrorConstraintFunctor_Equidistant, 2, 3, 6, 6>(new ResidualErrorConstraintFunctor_Equidistant(w, h, observation_first.homogeneous(), observation_second.homogeneous(), radius));
+    case EQUIDISTANT_CAMERA_RADIAL3:
+      return new ceres::AutoDiffCostFunction<ResidualErrorConstraintFunctor_EquidistantRadialK3, 2, 6, 6, 6>(new ResidualErrorConstraintFunctor_EquidistantRadialK3(w, h, observation_first.homogeneous(), observation_second.homogeneous(), radius));
+    default:
+      throw std::logic_error("Cannot create cost function, unrecognized intrinsic type in BA.");
+  } 
+}
+
+void BundleAdjustmentCeres::CeresOptions::setDenseBA()
 {
   // default configuration use a DENSE representation
   preconditionerType  = ceres::JACOBI;
   linearSolverType = ceres::DENSE_SCHUR;
   sparseLinearAlgebraLibraryType = ceres::SUITE_SPARSE; // not used but just to avoid a warning in ceres
-  ALICEVISION_LOG_DEBUG("BundleAdjustmentSymbolic[Ceres]: DENSE_SCHUR");
+  ALICEVISION_LOG_DEBUG("BundleAdjustment[Ceres]: DENSE_SCHUR");
 }
 
-void BundleAdjustmentSymbolicCeres::CeresOptions::setSparseBA()
+void BundleAdjustmentCeres::CeresOptions::setSparseBA()
 {
   preconditionerType = ceres::JACOBI;
   // if Sparse linear solver are available
@@ -93,30 +359,30 @@ void BundleAdjustmentSymbolicCeres::CeresOptions::setSparseBA()
   {
     sparseLinearAlgebraLibraryType = ceres::SUITE_SPARSE;
     linearSolverType = ceres::SPARSE_SCHUR;
-    ALICEVISION_LOG_DEBUG("BundleAdjustmentSymbolic[Ceres]: SPARSE_SCHUR, SUITE_SPARSE");
+    ALICEVISION_LOG_DEBUG("BundleAdjustment[Ceres]: SPARSE_SCHUR, SUITE_SPARSE");
   }
 #if ALICEVISION_CERES_HAS_CXSPARSE
   else if (ceres::IsSparseLinearAlgebraLibraryTypeAvailable(ceres::CX_SPARSE))
   {
     sparseLinearAlgebraLibraryType = ceres::CX_SPARSE;
     linearSolverType = ceres::SPARSE_SCHUR;
-    ALICEVISION_LOG_DEBUG("BundleAdjustmentSymbolic[Ceres]: SPARSE_SCHUR, CX_SPARSE");
+    ALICEVISION_LOG_DEBUG("BundleAdjustment[Ceres]: SPARSE_SCHUR, CX_SPARSE");
   }
 #endif
   else if (ceres::IsSparseLinearAlgebraLibraryTypeAvailable(ceres::EIGEN_SPARSE))
   {
     sparseLinearAlgebraLibraryType = ceres::EIGEN_SPARSE;
     linearSolverType = ceres::SPARSE_SCHUR;
-    ALICEVISION_LOG_DEBUG("BundleAdjustmentSymbolic[Ceres]: SPARSE_SCHUR, EIGEN_SPARSE");
+    ALICEVISION_LOG_DEBUG("BundleAdjustment[Ceres]: SPARSE_SCHUR, EIGEN_SPARSE");
   }
   else
   {
     linearSolverType = ceres::DENSE_SCHUR;
-    ALICEVISION_LOG_WARNING("BundleAdjustmentSymbolic[Ceres]: no sparse BA available, fallback to dense BA.");
+    ALICEVISION_LOG_WARNING("BundleAdjustment[Ceres]: no sparse BA available, fallback to dense BA.");
   }
 }
 
-bool BundleAdjustmentSymbolicCeres::Statistics::exportToFile(const std::string& folder, const std::string& filename) const
+bool BundleAdjustmentCeres::Statistics::exportToFile(const std::string& folder, const std::string& filename) const
 {
   std::ofstream os;
   os.open((fs::path(folder) / filename).string(), std::ios::app);
@@ -145,12 +411,8 @@ bool BundleAdjustmentSymbolicCeres::Statistics::exportToFile(const std::string& 
   std::size_t posesWithDistUpperThanTen = 0;
 
   for(const auto& it : nbCamerasPerDistance)
-  {
-      if (it.first >= 10)
-      {
-          posesWithDistUpperThanTen += it.second;
-      }
-  }
+    if (it.first >= 10)
+      posesWithDistUpperThanTen += it.second;
 
   os << time << ";"
      << states[EParameter::POSE][EParameterState::REFINED]  << ";"
@@ -183,7 +445,7 @@ bool BundleAdjustmentSymbolicCeres::Statistics::exportToFile(const std::string& 
   return true;
 }
 
-void BundleAdjustmentSymbolicCeres::Statistics::show() const
+void BundleAdjustmentCeres::Statistics::show() const
 {
   std::map<EParameter, std::map<EParameterState, std::size_t>> states = parametersStates;
   std::stringstream ss;
@@ -241,7 +503,7 @@ void BundleAdjustmentSymbolicCeres::Statistics::show() const
                         << "\t- final   RMSE: " << RMSEfinal);
 }
 
-void BundleAdjustmentSymbolicCeres::setSolverOptions(ceres::Solver::Options& solverOptions) const
+void BundleAdjustmentCeres::setSolverOptions(ceres::Solver::Options& solverOptions) const
 {
   solverOptions.preconditioner_type = _ceresOptions.preconditionerType;
   solverOptions.linear_solver_type = _ceresOptions.linearSolverType;
@@ -253,7 +515,7 @@ void BundleAdjustmentSymbolicCeres::setSolverOptions(ceres::Solver::Options& sol
   /*solverOptions.function_tolerance = 1e-12;
   solverOptions.gradient_tolerance = 1e-12;
   solverOptions.parameter_tolerance = 1e-12;*/
-
+  
 #if CERES_VERSION_MAJOR < 2
   solverOptions.num_linear_solver_threads = _ceresOptions.nbThreads;
 #endif
@@ -265,10 +527,68 @@ void BundleAdjustmentSymbolicCeres::setSolverOptions(ceres::Solver::Options& sol
   }
 }
 
-void BundleAdjustmentSymbolicCeres::addExtrinsicsToProblem(const sfmData::SfMData& sfmData, BundleAdjustment::ERefineOptions refineOptions, ceres::Problem& problem)
+void BundleAdjustmentCeres::addExtrinsicsToProblem(const sfmData::SfMData& sfmData, BundleAdjustment::ERefineOptions refineOptions, ceres::Problem& problem)
 {
   const bool refineTranslation = refineOptions & BundleAdjustment::REFINE_TRANSLATION;
   const bool refineRotation = refineOptions & BundleAdjustment::REFINE_ROTATION;
+
+  const auto addPose = [&](const sfmData::CameraPose& cameraPose, bool isConstant, std::array<double,6>& poseBlock)
+  {
+    const Mat3& R = cameraPose.getTransform().rotation();
+    const Vec3& t = cameraPose.getTransform().translation();
+
+    double angleAxis[3];
+    ceres::RotationMatrixToAngleAxis(static_cast<const double *>(R.data()), angleAxis);
+    poseBlock.at(0) = angleAxis[0];
+    poseBlock.at(1) = angleAxis[1];
+    poseBlock.at(2) = angleAxis[2];
+    poseBlock.at(3) = t(0);
+    poseBlock.at(4) = t(1);
+    poseBlock.at(5) = t(2);
+
+    double* poseBlockPtr = poseBlock.data();
+    problem.AddParameterBlock(poseBlockPtr, 6);
+
+    // add pose parameter to the all parameters blocks pointers list
+    _allParametersBlocks.push_back(poseBlockPtr);
+
+    // keep the camera extrinsics constants
+    if(cameraPose.isLocked() || isConstant || (!refineTranslation && !refineRotation))
+    {
+      // set the whole parameter block as constant.
+      _statistics.addState(EParameter::POSE, EParameterState::CONSTANT);
+      problem.SetParameterBlockConstant(poseBlockPtr);
+      return;
+    }
+
+    // constant parameters
+    std::vector<int> constantExtrinsic;
+
+    // don't refine rotations
+    if(!refineRotation)
+    {
+      constantExtrinsic.push_back(0);
+      constantExtrinsic.push_back(1);
+      constantExtrinsic.push_back(2);
+    }
+
+    // don't refine translations
+    if(!refineTranslation)
+    {
+      constantExtrinsic.push_back(3);
+      constantExtrinsic.push_back(4);
+      constantExtrinsic.push_back(5);
+    }
+
+    // subset parametrization
+    if(!constantExtrinsic.empty())
+    {
+      auto* subsetManifold = new ceres::SubsetManifold(6, constantExtrinsic);
+      problem.SetManifold(poseBlockPtr, subsetManifold);
+    }
+
+    _statistics.addState(EParameter::POSE, EParameterState::REFINED);
+  };
 
   // setup poses data
   for(const auto& posePair : sfmData.getPoses())
@@ -285,7 +605,7 @@ void BundleAdjustmentSymbolicCeres::addExtrinsicsToProblem(const sfmData::SfMDat
 
     const bool isConstant = (getPoseState(poseId) == EParameterState::CONSTANT);
 
-    addPose(pose, isConstant, _posesBlocks[poseId], problem, refineTranslation, refineRotation);
+    addPose(pose, isConstant, _posesBlocks[poseId]);
   }
 
   // setup sub-poses data
@@ -304,22 +624,17 @@ void BundleAdjustmentSymbolicCeres::addExtrinsicsToProblem(const sfmData::SfMDat
 
       const bool isConstant = (rigSubPose.status == sfmData::ERigSubPoseStatus::CONSTANT);
 
-      addPose(sfmData::CameraPose(rigSubPose.pose), isConstant, _rigBlocks[rigId][subPoseId], problem, refineTranslation, refineRotation);
+      addPose(sfmData::CameraPose(rigSubPose.pose), isConstant, _rigBlocks[rigId][subPoseId]);
     }
   }
-
-
-  //Add default rig pose
-  addPose(sfmData::CameraPose(), true, _rigNull, problem, refineTranslation, refineRotation);
 }
 
-void BundleAdjustmentSymbolicCeres::addIntrinsicsToProblem(const sfmData::SfMData& sfmData, BundleAdjustment::ERefineOptions refineOptions, ceres::Problem& problem)
+void BundleAdjustmentCeres::addIntrinsicsToProblem(const sfmData::SfMData& sfmData, BundleAdjustment::ERefineOptions refineOptions, ceres::Problem& problem)
 {
   const bool refineIntrinsicsOpticalCenter = (refineOptions & REFINE_INTRINSICS_OPTICALOFFSET_ALWAYS) || (refineOptions & REFINE_INTRINSICS_OPTICALOFFSET_IF_ENOUGH_DATA);
   const bool refineIntrinsicsFocalLength = refineOptions & REFINE_INTRINSICS_FOCAL;
   const bool refineIntrinsicsDistortion = refineOptions & REFINE_INTRINSICS_DISTORTION;
   const bool refineIntrinsics = refineIntrinsicsDistortion || refineIntrinsicsFocalLength || refineIntrinsicsOpticalCenter;
-  const bool fixFocalRatio = true;
 
   std::map<IndexT, std::size_t> intrinsicsUsage;
 
@@ -354,12 +669,11 @@ void BundleAdjustmentSymbolicCeres::addIntrinsicsToProblem(const sfmData::SfMDat
 
     assert(isValid(intrinsicPtr->getType()));
 
-    _intrinsicObjects[intrinsicId].reset(intrinsicPtr->clone());
-
     std::vector<double>& intrinsicBlock = _intrinsicsBlocks[intrinsicId];
     intrinsicBlock = intrinsicPtr->getParams();
 
     double* intrinsicBlockPtr = intrinsicBlock.data();
+
     problem.AddParameterBlock(intrinsicBlockPtr, intrinsicBlock.size());
 
     // add intrinsic parameter to the all parameters blocks pointers list
@@ -385,7 +699,7 @@ void BundleAdjustmentSymbolicCeres::addIntrinsicsToProblem(const sfmData::SfMDat
     if(refineIntrinsicsFocalLength)
     {
       std::shared_ptr<camera::IntrinsicScaleOffset> intrinsicScaleOffset = std::dynamic_pointer_cast<camera::IntrinsicScaleOffset>(intrinsicPtr);
-      if(intrinsicScaleOffset->getInitialScale().x() > 0 && intrinsicScaleOffset->getInitialScale().y() > 0 && _ceresOptions.useFocalPrior)
+      if (intrinsicScaleOffset->getInitialScale().x() > 0 && intrinsicScaleOffset->getInitialScale().y() > 0)
       {
         // if we have an initial guess, we only authorize a margin around this value.
         assert(intrinsicBlock.size() >= 1);
@@ -403,7 +717,7 @@ void BundleAdjustmentSymbolicCeres::addIntrinsicsToProblem(const sfmData::SfMDat
         problem.SetParameterLowerBound(intrinsicBlockPtr, 1, 0.0);
       }
 
-      focalRatio = intrinsicBlockPtr[1] / intrinsicBlockPtr[0];
+      focalRatio = intrinsicBlockPtr[0] / intrinsicBlockPtr[1];
 
       std::shared_ptr<camera::IntrinsicScaleOffset> castedcam_iso = std::dynamic_pointer_cast<camera::IntrinsicScaleOffset>(intrinsicPtr);
       if (castedcam_iso)
@@ -417,11 +731,12 @@ void BundleAdjustmentSymbolicCeres::addIntrinsicsToProblem(const sfmData::SfMDat
       lockFocal = true;
     }
 
-    const bool optional_center = ((refineOptions & REFINE_INTRINSICS_OPTICALOFFSET_IF_ENOUGH_DATA) && (usageCount > _minNbImagesToRefineOpticalCenter));
-    if ((refineOptions & REFINE_INTRINSICS_OPTICALOFFSET_ALWAYS) || optional_center)
+    // optical center
+    if((refineOptions & REFINE_INTRINSICS_OPTICALOFFSET_ALWAYS) ||
+       ((refineOptions & REFINE_INTRINSICS_OPTICALOFFSET_IF_ENOUGH_DATA) && _minNbImagesToRefineOpticalCenter > 0 && usageCount >= _minNbImagesToRefineOpticalCenter))
     {
       // refine optical center within 10% of the image size.
-      assert(intrinsicBlock.size() >= 4);
+      assert(intrinsicBlock.size() >= 3);
 
       const double opticalCenterMinPercent = -0.05;
       const double opticalCenterMaxPercent =  0.05;
@@ -439,19 +754,22 @@ void BundleAdjustmentSymbolicCeres::addIntrinsicsToProblem(const sfmData::SfMDat
     }
 
     // lens distortion
-    if(!refineIntrinsicsDistortion)
+    if (!refineIntrinsicsDistortion ||
+        intrinsicPtr->getDistortionInitializationMode() == camera::EInitMode::CALIBRATED)
     {
       lockDistortion = true;
     }
 
-    IntrinsicsManifoldSymbolic* subsetManifold = new IntrinsicsManifoldSymbolic(intrinsicBlock.size(), focalRatio,
+    
+    IntrinsicsManifold* subsetManifold = new IntrinsicsManifold(intrinsicBlock.size(), focalRatio,
                                                                 lockFocal, lockRatio, lockCenter, lockDistortion);
     problem.SetManifold(intrinsicBlockPtr, subsetManifold);
+
     _statistics.addState(EParameter::INTRINSIC, EParameterState::REFINED);
   }
 }
 
-void BundleAdjustmentSymbolicCeres::addLandmarksToProblem(const sfmData::SfMData& sfmData, ERefineOptions refineOptions, ceres::Problem& problem)
+void BundleAdjustmentCeres::addLandmarksToProblem(const sfmData::SfMData& sfmData, ERefineOptions refineOptions, ceres::Problem& problem)
 {
   const bool refineStructure = refineOptions & REFINE_STRUCTURE;
 
@@ -488,9 +806,6 @@ void BundleAdjustmentSymbolicCeres::addLandmarksToProblem(const sfmData::SfMData
       const sfmData::View& view = sfmData.getView(observationPair.first);
       const sfmData::Observation& observation = observationPair.second;
 
-      //Get shared object clone
-      const std::shared_ptr<IntrinsicBase> intrinsic = _intrinsicObjects[view.getIntrinsicId()];
-
       // each residual block takes a point and a camera as input and outputs a 2
       // dimensional residual. Internally, the cost function stores the observed
       // image location and compares the reprojection against the observation.
@@ -502,15 +817,6 @@ void BundleAdjustmentSymbolicCeres::addLandmarksToProblem(const sfmData::SfMData
       double* poseBlockPtr = _posesBlocks.at(view.getPoseId()).data();
       double* intrinsicBlockPtr = _intrinsicsBlocks.at(view.getIntrinsicId()).data();
 
-      bool withRig = (view.isPartOfRig() && !view.isPoseIndependant());
-      double * rigBlockPtr = nullptr;
-      if (withRig) {
-        rigBlockPtr = _rigBlocks.at(view.getRigId()).at(view.getSubPoseId()).data();
-      }
-      else {
-        rigBlockPtr = _rigNull.data();
-      }
-
       // apply a specific parameter ordering:
       if(_ceresOptions.useParametersOrdering)
       {
@@ -519,15 +825,29 @@ void BundleAdjustmentSymbolicCeres::addLandmarksToProblem(const sfmData::SfMData
         _linearSolverOrdering.AddElementToGroup(intrinsicBlockPtr, 2);
       }
 
-      if (withRig)
+      if(view.isPartOfRig() && !view.isPoseIndependant())
       {
-        ceres::CostFunction* costFunction = new CostProjection(observation, intrinsic, withRig);
-        problem.AddResidualBlock(costFunction, lossFunction, poseBlockPtr, rigBlockPtr, intrinsicBlockPtr, landmarkBlockPtr);
+        ceres::CostFunction* costFunction = createRigCostFunctionFromIntrinsics(sfmData.getIntrinsicPtr(view.getIntrinsicId()), observation);
+
+        double* rigBlockPtr = _rigBlocks.at(view.getRigId()).at(view.getSubPoseId()).data();
+        _linearSolverOrdering.AddElementToGroup(rigBlockPtr, 1);
+
+        problem.AddResidualBlock(costFunction,
+            lossFunction,
+            intrinsicBlockPtr,
+            poseBlockPtr,
+            rigBlockPtr, // subpose of the cameras rig
+            landmarkBlockPtr); // do we need to copy 3D point to avoid false motion, if failure ?
       }
       else
       {
-        ceres::CostFunction* costFunction = new CostProjectionSimple(observation, intrinsic);
-        problem.AddResidualBlock(costFunction, lossFunction, poseBlockPtr, intrinsicBlockPtr, landmarkBlockPtr);
+        ceres::CostFunction* costFunction = createCostFunctionFromIntrinsics(sfmData.getIntrinsicPtr(view.getIntrinsicId()), observation);
+
+        problem.AddResidualBlock(costFunction,
+            lossFunction,
+            intrinsicBlockPtr,
+            poseBlockPtr,
+            landmarkBlockPtr); //do we need to copy 3D point to avoid false motion, if failure ?
       }
 
       if(!refineStructure || getLandmarkState(landmarkId) == EParameterState::CONSTANT)
@@ -544,11 +864,11 @@ void BundleAdjustmentSymbolicCeres::addLandmarksToProblem(const sfmData::SfMData
   }
 }
 
-void BundleAdjustmentSymbolicCeres::addConstraints2DToProblem(const sfmData::SfMData& sfmData, ERefineOptions refineOptions, ceres::Problem& problem)
+void BundleAdjustmentCeres::addConstraints2DToProblem(const sfmData::SfMData& sfmData, ERefineOptions refineOptions, ceres::Problem& problem)
 {
   // set a LossFunction to be less penalized by false measurements.
   // note: set it to NULL if you don't want use a lossFunction.
-  ceres::LossFunction* lossFunction = new ceres::HuberLoss(Square(8.0)); // TODO: make the LOSS function and the parameter an option
+  ceres::LossFunction* lossFunction = _ceresOptions.lossFunction.get();
 
   for (const auto & constraint : sfmData.getConstraints2D()) {
     const sfmData::View& view_1 = sfmData.getView(constraint.ViewFirst);
@@ -559,53 +879,28 @@ void BundleAdjustmentSymbolicCeres::addConstraints2DToProblem(const sfmData::SfM
     assert(getPoseState(view_2.getPoseId()) != EParameterState::IGNORED);
     assert(getIntrinsicState(view_2.getIntrinsicId()) != EParameterState::IGNORED);
 
-    /* Get pose */
     double * poseBlockPtr_1 = _posesBlocks.at(view_1.getPoseId()).data();
     double * poseBlockPtr_2 = _posesBlocks.at(view_2.getPoseId()).data();
 
-    /*Get intrinsics */
     double * intrinsicBlockPtr_1 = _intrinsicsBlocks.at(view_1.getIntrinsicId()).data();
     double * intrinsicBlockPtr_2 = _intrinsicsBlocks.at(view_2.getIntrinsicId()).data();
 
-    /* For the moment assume a unique camera */
+    //For the moment assume a unique camera
     assert(intrinsicBlockPtr_1 == intrinsicBlockPtr_2);
 
-    std::shared_ptr<IntrinsicBase> intrinsic = sfmData.getIntrinsicsharedPtr(view_1.getIntrinsicId());
-    std::shared_ptr<camera::Equidistant> equidistant = std::dynamic_pointer_cast<camera::Equidistant>(intrinsic);
-    std::shared_ptr<camera::Pinhole> pinhole = std::dynamic_pointer_cast<camera::Pinhole>(intrinsic);
 
-    if (equidistant != nullptr)  
-    {
-      ceres::CostFunction* costFunction = new CostPanoramaEquidistant(constraint.ObservationFirst.x, constraint.ObservationSecond.x, equidistant);
-      problem.AddResidualBlock(costFunction, lossFunction, poseBlockPtr_1, poseBlockPtr_2, intrinsicBlockPtr_1);
-
-      /* Symmetry */
-      costFunction = new CostPanoramaEquidistant(constraint.ObservationSecond.x, constraint.ObservationFirst.x, equidistant);
-      problem.AddResidualBlock(costFunction, lossFunction, poseBlockPtr_2, poseBlockPtr_1, intrinsicBlockPtr_1);
-    }
-    else if (pinhole != nullptr)  
-    {
-      ceres::CostFunction* costFunction = new CostPanoramaPinHole(constraint.ObservationFirst.x, constraint.ObservationSecond.x, pinhole);
-      problem.AddResidualBlock(costFunction, lossFunction, poseBlockPtr_1, poseBlockPtr_2, intrinsicBlockPtr_1);
-      /* Symmetry */
-      costFunction = new CostPanoramaPinHole(constraint.ObservationSecond.x, constraint.ObservationFirst.x, pinhole);
-      problem.AddResidualBlock(costFunction, lossFunction, poseBlockPtr_2, poseBlockPtr_1, intrinsicBlockPtr_1);
-    }
-    else {
-      ALICEVISION_LOG_ERROR("Incompatible camera for a 2D constraint");
-      return;
-    }
+    ceres::CostFunction* costFunction = createConstraintsCostFunctionFromIntrinsics(sfmData.getIntrinsicPtr(view_1.getIntrinsicId()), constraint.ObservationFirst.x, constraint.ObservationSecond.x);
+    problem.AddResidualBlock(costFunction, lossFunction, intrinsicBlockPtr_1, poseBlockPtr_1, poseBlockPtr_2);
   }
 }
 
-void BundleAdjustmentSymbolicCeres::addRotationPriorsToProblem(const sfmData::SfMData& sfmData, ERefineOptions refineOptions, ceres::Problem& problem)
+void BundleAdjustmentCeres::addRotationPriorsToProblem(const sfmData::SfMData& sfmData, ERefineOptions refineOptions, ceres::Problem& problem)
 {
   // set a LossFunction to be less penalized by false measurements.
   // note: set it to NULL if you don't want use a lossFunction.
   ceres::LossFunction* lossFunction = nullptr;
 
   for (const auto & prior : sfmData.getRotationPriors()) {
-
     const sfmData::View& view_1 = sfmData.getView(prior.ViewFirst);
     const sfmData::View& view_2 = sfmData.getView(prior.ViewSecond);
 
@@ -615,12 +910,13 @@ void BundleAdjustmentSymbolicCeres::addRotationPriorsToProblem(const sfmData::Sf
     double * poseBlockPtr_1 = _posesBlocks.at(view_1.getPoseId()).data();
     double * poseBlockPtr_2 = _posesBlocks.at(view_2.getPoseId()).data();
 
-    ceres::CostFunction* costFunction = new CostRotationPrior(prior._second_R_first);
+
+    ceres::CostFunction* costFunction = new ceres::AutoDiffCostFunction<ResidualErrorRotationPriorFunctor, 3, 6, 6>(new ResidualErrorRotationPriorFunctor(prior._second_R_first));
     problem.AddResidualBlock(costFunction, lossFunction, poseBlockPtr_1, poseBlockPtr_2);
   }
 }
 
-void BundleAdjustmentSymbolicCeres::createProblem(const sfmData::SfMData& sfmData,
+void BundleAdjustmentCeres::createProblem(const sfmData::SfMData& sfmData,
                                           ERefineOptions refineOptions,
                                           ceres::Problem& problem)
 {
@@ -647,7 +943,20 @@ void BundleAdjustmentSymbolicCeres::createProblem(const sfmData::SfMData& sfmDat
   addRotationPriorsToProblem(sfmData, refineOptions, problem);
 }
 
-void BundleAdjustmentSymbolicCeres::updateFromSolution(sfmData::SfMData& sfmData, ERefineOptions refineOptions) const
+ void BundleAdjustmentCeres::resetProblem()
+{
+  _statistics = Statistics();
+
+  _allParametersBlocks.clear();
+  _posesBlocks.clear();
+  _intrinsicsBlocks.clear();
+  _landmarksBlocks.clear();
+  _rigBlocks.clear();
+
+  _linearSolverOrdering.Clear();
+}
+
+void BundleAdjustmentCeres::updateFromSolution(sfmData::SfMData& sfmData, ERefineOptions refineOptions) const
 {
   const bool refinePoses = (refineOptions & REFINE_ROTATION) || (refineOptions & REFINE_TRANSLATION);
   const bool refineIntrinsicsOpticalCenter = (refineOptions & REFINE_INTRINSICS_OPTICALOFFSET_ALWAYS) || (refineOptions & REFINE_INTRINSICS_OPTICALOFFSET_IF_ENOUGH_DATA);
@@ -666,10 +975,14 @@ void BundleAdjustmentSymbolicCeres::updateFromSolution(sfmData::SfMData& sfmData
       if(getPoseState(poseId) != EParameterState::REFINED)
         continue;
 
-      const SE3::Matrix & poseBlock = _posesBlocks.at(poseId);
+      const std::array<double,6>& poseBlock = _posesBlocks.at(poseId);
+
+      Mat3 R_refined;
+      ceres::AngleAxisToRotationMatrix(poseBlock.data(), R_refined.data());
+      const Vec3 t_refined(poseBlock.at(3), poseBlock.at(4), poseBlock.at(5));
 
       // update the pose
-      posePair.second.setTransform(poseFromRT(poseBlock.block<3, 3>(0, 0), poseBlock.block<3, 1>(0, 3)));
+      posePair.second.setTransform(poseFromRT(R_refined, t_refined));
     }
 
     // rig sub-poses
@@ -680,10 +993,14 @@ void BundleAdjustmentSymbolicCeres::updateFromSolution(sfmData::SfMData& sfmData
       for(const auto& subPoseit : rigIt.second)
       {
         sfmData::RigSubPose& subPose = rig.getSubPose(subPoseit.first);
-        const SE3::Matrix & subPoseBlock = subPoseit.second;
+        const std::array<double,6>& subPoseBlock = subPoseit.second;
+
+        Mat3 R_refined;
+        ceres::AngleAxisToRotationMatrix(subPoseBlock.data(), R_refined.data());
+        const Vec3 t_refined(subPoseBlock.at(3), subPoseBlock.at(4), subPoseBlock.at(5));
 
         // update the sub-pose
-        subPose.pose = poseFromRT(subPoseBlock.block<3, 3>(0, 0), subPoseBlock.block<3, 1>(0, 3));
+        subPose.pose = poseFromRT(R_refined, t_refined);
       }
     }
   }
@@ -712,18 +1029,16 @@ void BundleAdjustmentSymbolicCeres::updateFromSolution(sfmData::SfMData& sfmData
       sfmData::Landmark& landmark = sfmData.getLandmarks().at(landmarkId);
 
       // do not update a camera pose set as Ignored or Constant in the Local strategy
-      if(getLandmarkState(landmarkId) != EParameterState::REFINED) {
+      if(getLandmarkState(landmarkId) != EParameterState::REFINED)
         continue;
-      }
 
-      for(std::size_t i = 0; i < 3; ++i) {
+      for(std::size_t i = 0; i < 3; ++i)
         landmark.X(Eigen::Index(i))= landmarksBlockPair.second.at(i);
-      }
     }
   }
 }
 
-void BundleAdjustmentSymbolicCeres::createJacobian(const sfmData::SfMData& sfmData,
+void BundleAdjustmentCeres::createJacobian(const sfmData::SfMData& sfmData,
                                            ERefineOptions refineOptions,
                                            ceres::CRSMatrix& jacobian)
 {
@@ -744,13 +1059,11 @@ void BundleAdjustmentSymbolicCeres::createJacobian(const sfmData::SfMData& sfmDa
   problem.Evaluate(evalOpt, &cost, NULL, NULL, &jacobian);
 }
 
-bool BundleAdjustmentSymbolicCeres::adjust(sfmData::SfMData& sfmData, ERefineOptions refineOptions)
+bool BundleAdjustmentCeres::adjust(sfmData::SfMData& sfmData, ERefineOptions refineOptions)
 {
   // create problem
   ceres::Problem::Options problemOptions;
   problemOptions.loss_function_ownership = ceres::DO_NOT_TAKE_OWNERSHIP;
-  problemOptions.evaluation_callback = this;
-
   ceres::Problem problem(problemOptions);
   createProblem(sfmData, refineOptions, problem);
 
@@ -764,9 +1077,8 @@ bool BundleAdjustmentSymbolicCeres::adjust(sfmData::SfMData& sfmData, ERefineOpt
   ceres::Solve(options, &problem, &summary);
 
   // print summary
-  if(_ceresOptions.summary) {
+  if(_ceresOptions.summary)
     ALICEVISION_LOG_INFO(summary.FullReport());
-  }
 
   // solution is not usable
   if(!summary.IsSolutionUsable())
@@ -774,9 +1086,10 @@ bool BundleAdjustmentSymbolicCeres::adjust(sfmData::SfMData& sfmData, ERefineOpt
     ALICEVISION_LOG_WARNING("Bundle Adjustment failed, the solution is not usable.");
     return false;
   }
-
+  
   // update input sfmData with the solution
   updateFromSolution(sfmData, refineOptions);
+
 
   // store some statitics from the summary
   _statistics.time = summary.total_time_in_seconds;
@@ -788,22 +1101,9 @@ bool BundleAdjustmentSymbolicCeres::adjust(sfmData::SfMData& sfmData, ERefineOpt
 
   //store distance histogram for local strategy
   if(useLocalStrategy())
-  {
     _statistics.nbCamerasPerDistance = _localGraph->getDistancesHistogram();
-  }
 
   return true;
-}
-
-void BundleAdjustmentSymbolicCeres::PrepareForEvaluation(bool evaluate_jacobians, bool new_evaluation_point)
-{
-    if (new_evaluation_point)
-    {
-        for (const auto & pair : _intrinsicsBlocks)
-        {
-            _intrinsicObjects[pair.first]->updateFromParams(pair.second);
-        }
-    }
 }
 
 } // namespace sfm
