+++ src/aliceVision/sfm/bundle/bundleAdjustment_test.cpp
