+++ src/aliceVision/sfm/bundle/costfunctions/panoramaEquidistant.hpp
