+++ src/aliceVision/sfm/bundle/costfunctions/panoramaPinhole.hpp
