+++ src/aliceVision/sfm/bundle/costfunctions/projection.hpp
