+++ src/aliceVision/sfm/bundle/costfunctions/projectionSimple.hpp
