+++ src/aliceVision/sfm/bundle/costfunctions/rotationPrior.hpp
