+++ src/aliceVision/sfm/bundle/manifolds/intrinsics.hpp
