+++ src/aliceVision/sfm/bundle/manifolds/se3.hpp
