+++ src/aliceVision/sfm/bundle/manifolds/so2.hpp
