+++ src/aliceVision/sfm/bundle/manifolds/so3.hpp
